brutalizer/pyproject.toml
[tool.poetry]
name = "brutalizer"
version = "0.1.0"
description = ""
authors = ["Justin Litchfield <litch@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.50.0"
grpcio-tools = "^1.50.0"
googleapis-common-protos = "^1.56.4"
flask = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
